Groundwork for Projectiles (very rough)
sorry for the slack

Began work on the "projectiles" class, can and will be used for both allied and enemy purposes.

Also uploaded a sprite sheet for use on the main repository, has a bunch of options for use. 10 are currently defined in a dictionary albeit unused in the code.

Added a movement option to press [Left Shift], slows down movement halfway for precision.